# extracted_api/docs/openapi-original-fastapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecuador Judicial API",
    "description": "API comercial para consultas puntuales de procesos judiciales de Ecuador.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/juicios/{identificacion}": {
      "get": {
        "summary": "Api Juicios",
        "operationId": "api_juicios_api_juicios__identificacion__get",
        "parameters": [
          {
            "name": "identificacion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identificacion"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Juicio"
                  },
                  "title": "Response Api Juicios Api Juicios  Identificacion  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/buscar": {
      "post": {
        "summary": "Api V1 Buscar Causas",
        "operationId": "api_v1_buscar_causas_api_v1_causas_buscar_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuscarCausasRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Buscar Causas Api V1 Causas Buscar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/actuaciones": {
      "get": {
        "summary": "Api V1 Causa Actuaciones",
        "operationId": "api_v1_causa_actuaciones_api_v1_causas__id_juicio__actuaciones_get",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Causa Actuaciones Api V1 Causas  Id Juicio  Actuaciones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/actuaciones/paginadas": {
      "get": {
        "summary": "Api V1 Causa Actuaciones Paginadas",
        "operationId": "api_v1_causa_actuaciones_paginadas_api_v1_causas__id_juicio__actuaciones_paginadas_get",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "desc",
              "title": "Orden"
            }
          },
          {
            "name": "fechaDesde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fechadesde"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fechahasta"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "name": "tieneDocumento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tienedocumento"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Causa Actuaciones Paginadas Api V1 Causas  Id Juicio  Actuaciones Paginadas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/resoluciones": {
      "get": {
        "summary": "Api V1 Causa Resoluciones",
        "operationId": "api_v1_causa_resoluciones_api_v1_causas__id_juicio__resoluciones_get",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "fechaDesde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fechadesde"
            }
          },
          {
            "name": "fechaHasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fechahasta"
            }
          },
          {
            "name": "incluirSinDocumento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Incluirsindocumento"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Causa Resoluciones Api V1 Causas  Id Juicio  Resoluciones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/actuaciones/{codigo_actuacion}/documentos": {
      "get": {
        "summary": "Api V1 Causa Actuacion Documentos",
        "operationId": "api_v1_causa_actuacion_documentos_api_v1_causas__id_juicio__actuaciones__codigo_actuacion__documentos_get",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "codigo_actuacion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo Actuacion"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Causa Actuacion Documentos Api V1 Causas  Id Juicio  Actuaciones  Codigo Actuacion  Documentos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/abandono/riesgo": {
      "get": {
        "summary": "Api V1 Causa Abandono Riesgo",
        "operationId": "api_v1_causa_abandono_riesgo_api_v1_causas__id_juicio__abandono_riesgo_get",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "fechaCorte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fechacorte"
            }
          },
          {
            "name": "alertaDias",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 180,
              "minimum": 1,
              "default": 30,
              "title": "Alertadias"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api V1 Causa Abandono Riesgo Api V1 Causas  Id Juicio  Abandono Riesgo Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/causas/{id_juicio}/pdf": {
      "post": {
        "summary": "Api V1 Causa Pdf",
        "operationId": "api_v1_causa_pdf_api_v1_causas__id_juicio__pdf_post",
        "parameters": [
          {
            "name": "id_juicio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id Juicio"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Juicio": {
        "properties": {
          "id_juicio": {
            "type": "string",
            "title": "Id Juicio",
            "examples": [
              "01283-2024-00123"
            ]
          },
          "numero_proceso": {
            "type": "string",
            "title": "Numero Proceso"
          },
          "judicatura": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judicatura"
          },
          "materia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materia"
          },
          "accion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accion"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "demandado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demandado"
          },
          "fecha_ingreso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Ingreso"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado"
          },
          "fuente": {
            "type": "string",
            "title": "Fuente",
            "default": "funcion_judicial_ec"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "id_juicio",
          "numero_proceso"
        ],
        "title": "Juicio"
      },
...
    }
  }
}